Added more proofs
