+++ .cbmc-batch/jobs/array_list_init_static/array_list_init_static.yaml
@@ -1,0 +1,7 @@
+cbmcpkg: cbmc-memcpy-loop.tar.gz
+jobos: ubuntu16
+cbmcflags: "--bounds-check;--pointer-check;--function;aws_array_list_init_static_verify"
+goto: proofs.goto
+expected: "SUCCESSFUL"
+coverage: false
+report: false
