+++ .cbmc-batch/jobs/array_list_set_at/array_list_set_at.yaml
@@ -3,6 +3,5 @@ jobos: ubuntu16
 cbmcflags: "--bounds-check;--pointer-check;--function;aws_array_list_set_at_verify;--unwindset;strlen.0:33,memcpy.0:33,__builtin___memcpy_chk.0:33,aws_mem_release:1;--unwinding-assertions"
 goto: proofs.goto
 expected: "SUCCESSFUL"
-property: false
 coverage: false
 report: false
